+++ CHANGELOG.md
@@ -19,6 +19,14 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
 
 # Version 1
 
+# 1.34.0 - 2024-06-17
+
+- Support to configure the throttle plugin.
+
+# 1.33.1 - 2024-05-27
+
+- Fixed extension to depend on the DependencyInjection component rather than the HttpKernel.
+
 # 1.33.0 - 2024-02-27
 
 - Support php-http/cache-plugin 2.0 and bump minimal version to 1.7 by defaulting the stream factory for cache to `httplug.psr17_stream_factory` (#448).